# docs/curieri/EPACKET-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "e-packet API",
    "version": "1.0",
...
  },
  "servers": [
    {
      "url": "https://zbdnzolswscjoxhpsbtt.supabase.co/functions/v1/api/v1"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
  "paths": {
...
    "/status": {
      "get": {
        "tags": [
          "AWB"
        ],
        "summary": "Statusul livrării",
        "description": "Statusul unui AWB din contul dumneavoastră. Un status mai nou de 30 de minute nu este\ncerut din nou curierului; un status final (livrat, returnat, …) nu se mai schimbă.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/awb"
          }
        ],
        "responses": {
          "200": {
            "description": "Statusul.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                },
                "example": {
                  "awb": "80512345678",
                  "courier": "DPD",
                  "status": "in_tranzit",
                  "label": "În tranzit",
                  "is_final": false,
                  "status_at": "2026-10-05T11:20:00Z",
                  "checked_at": "2026-10-05T12:01:00Z"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cheia API, `epk_live_…` sau `epk_test_…`."
      }
    },
    "parameters": {
...
      "awb": {
        "name": "awb",
        "in": "query",
        "required": true,
        "description": "Numărul AWB.",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Eroare.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string",
                "description": "În română."
              },
              "field": {
                "type": "string",
                "description": "Doar pentru erorile de câmp."
              }
            }
          }
        }
      },
...
      "Status": {
        "type": "object",
        "properties": {
          "awb": {
            "type": "string"
          },
          "courier": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Codul statusului e-packet."
          },
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Denumirea în română."
          },
          "is_final": {
            "type": "boolean"
          },
          "status_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "checked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      }
    }
  }
}